# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Face API",
    "description": "API para validação de identidade através de comparação facial entre uma selfie e um documento.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor local"
    }
  ],
  "tags": [
    {
      "name": "documents",
      "description": "Recursos relacionados à análise e validação de documentos"
    }
  ],
  "paths": {
    "/api/documents/analize": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Analisar documento e validar identidade",
        "description": "Recebe uma selfie e uma imagem de documento de identificação, realiza a comparação facial entre as imagens e retorna o resultado da validação de identidade.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "selfie",
                  "documento"
                ],
                "properties": {
                  "selfie": {
                    "type": "string",
                    "format": "binary",
                    "description": "Imagem contendo a selfie do usuário."
                  },
                  "documento": {
                    "type": "string",
                    "format": "binary",
                    "description": "Imagem do documento de identificação contendo o rosto do usuário."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validação processada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityValidationResponse"
                },
                "examples": {
                  "approved": {
                    "summary": "Identidade validada",
                    "value": {
                      "status": true,
                      "nivel": "alto",
                      "sugestao": "Identidade validada",
                      "distanciaCartografica": 0.42
                    }
                  },
                  "rejected": {
                    "summary": "Identidade não confirmada",
                    "value": {
                      "status": false,
                      "nivel": "baixo",
                      "sugestao": "Não foi possível confirmar a identidade, solicitar novos documentos",
                      "distanciaCartografica": 0.87
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "missingFiles": {
                    "summary": "Arquivos não enviados",
                    "value": {
                      "error": "Envie selfie e documento"
                    }
                  },
                  "invalidDistance": {
                    "summary": "Distância cartográfica inválida",
                    "value": {
                      "success": false,
                      "status": 400,
                      "error": "Distância cartográfica inválida"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Não foi possível realizar a comparação facial devido às imagens fornecidas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "noFaceDetected": {
                    "summary": "Rosto não detectado",
                    "value": {
                      "success": false,
                      "status": 422,
                      "error": "Não detectamos um rosto claro na imagem. Tente novamente."
                    }
                  },
                  "multipleFaces": {
                    "summary": "Múltiplos rostos detectados",
                    "value": {
                      "success": false,
                      "status": 422,
                      "error": "Mais de um rosto detectado. Envie uma foto individual."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno durante o processamento da validação.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "success": false,
                  "status": 500,
                  "error": "Erro ao processar validação"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "IdentityValidationResponse": {
        "type": "object",
        "required": [
          "status",
          "nivel",
          "sugestao",
          "distanciaCartografica"
        ],
        "properties": {
          "status": {
            "type": "boolean",
            "description": "Indica se a identidade foi validada.",
            "example": true
          },
          "nivel": {
            "type": "string",
            "description": "Nível de confiança definido pela análise da distância facial.",
            "example": "alto"
          },
          "sugestao": {
            "type": "string",
            "description": "Orientação baseada no resultado da validação.",
            "example": "Identidade validada"
          },
          "distanciaCartografica": {
            "type": "number",
            "format": "double",
            "description": "Distância calculada entre as representações faciais da selfie e do documento.",
            "example": 0.42
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica se a operação foi realizada com sucesso.",
            "example": false
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP retornado pela API.",
            "example": 422
          },
          "error": {
            "type": "string",
            "description": "Mensagem descrevendo o erro ocorrido.",
            "example": "Não detectamos um rosto claro na imagem. Tente novamente."
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}